refactor: Organize frontend
