--- public/js/game/index.json
@@ -1,121 +1,0 @@
-{
-  "blocks": [
-    { "type": "GrassBlock", "x": 0, "y": 140 },
-    { "type": "GrassBlock", "x": 100, "y": 140 },
-    { "type": "GrassBlock", "x": 200, "y": 140 },
-    { "type": "GrassBlock", "x": 300, "y": 240 },
-    { "type": "GrassBlock", "x": 400, "y": 340 },
-    { "type": "GrassBlock", "x": 500, "y": 340 },
-    { "type": "GrassBlock", "x": 500, "y": 440 },
-    { "type": "GrassBlock", "x": 2200, "y": 340 },
-    { "type": "GrassBlock", "x": 2400, "y": 340 },
-    { "type": "GrassBlock", "x": 0, "y": 240 },
-    { "type": "GrassBlock", "x": 100, "y": 240 },
-    { "type": "GrassBlock", "x": 200, "y": 240 },
-    { "type": "GrassBlock", "x": 200, "y": 340 },
-    { "type": "GrassBlock", "x": 300, "y": 340 },
-    { "type": "GrassBlock", "x": 400, "y": 440 },
-    { "type": "GrassBlock", "x": 300, "y": 440 },
-    { "type": "GrassBlock", "x": 200, "y": 440 },
-    { "type": "GrassBlock", "x": 100, "y": 340 },
-    { "type": "GrassBlock", "x": 100, "y": 440 },
-    { "type": "GrassBlock", "x": 0, "y": 340 },
-    { "type": "GrassBlock", "x": 0, "y": 440 },
-    { "type": "GrassBlock", "x": 400, "y": 540 },
-    { "type": "GrassBlock", "x": 300, "y": 540 },
-    { "type": "GrassBlock", "x": 200, "y": 540 },
-    { "type": "GrassBlock", "x": 100, "y": 540 },
-    { "type": "GrassBlock", "x": 0, "y": 640 },
-    { "type": "GrassBlock", "x": 0, "y": 540 },
-    { "type": "GrassBlock", "x": 100, "y": 640 },
-    { "type": "GrassBlock", "x": 200, "y": 640 },
-    { "type": "GrassBlock", "x": 300, "y": 640 },
-    { "type": "GrassBlock", "x": 200, "y": 740 },
-    { "type": "GrassBlock", "x": 0, "y": 740 },
-    { "type": "GrassBlock", "x": 0, "y": 840 },
-    { "type": "GrassBlock", "x": 100, "y": 840 },
-    { "type": "GrassBlock", "x": 100, "y": 740 },
-    { "type": "GrassBlock", "x": 2300, "y": 440 },
-    { "type": "GrassBlock", "x": 2400, "y": 440 },
-    { "type": "GrassBlock", "x": 2400, "y": 540 },
-    { "type": "GrassBlock", "x": 2300, "y": 540 },
-    { "type": "GrassBlock", "x": 2200, "y": 540 },
-    { "type": "GrassBlock", "x": 2400, "y": 640 },
-    { "type": "GrassBlock", "x": 2400, "y": 740 },
-    { "type": "GrassBlock", "x": 2400, "y": 840 },
-    { "type": "GrassBlock", "x": 2200, "y": 440 },
-    { "type": "GrassBlock", "x": 2300, "y": 240 },
-    { "type": "GrassBlock", "x": 2400, "y": 240 },
-    { "type": "GrassBlock", "x": 2200, "y": 240 },
-    { "type": "GrassBlock", "x": 2100, "y": 340 },
-    { "type": "GrassBlock", "x": 2000, "y": 340 },
-    { "type": "GrassBlock", "x": 1900, "y": 340 },
-    { "type": "GrassBlock", "x": 1800, "y": 340 },
-    { "type": "GrassBlock", "x": 1700, "y": 340 },
-    { "type": "GrassBlock", "x": 1500, "y": 340 },
-    { "type": "GrassBlock", "x": 1600, "y": 340 },
-    { "type": "GrassBlock", "x": 1400, "y": 340 },
-    { "type": "GrassBlock", "x": 1300, "y": 340 },
-    { "type": "GrassBlock", "x": 1200, "y": 340 },
-    { "type": "GrassBlock", "x": 1100, "y": 340 },
-    { "type": "GrassBlock", "x": 1000, "y": 340 },
-    { "type": "GrassBlock", "x": 900, "y": 340 },
-    { "type": "GrassBlock", "x": 800, "y": 340 },
-    { "type": "GrassBlock", "x": 600, "y": 340 },
-    { "type": "GrassBlock", "x": 700, "y": 340 },
-    { "type": "GrassBlock", "x": 800, "y": 840 },
-    { "type": "GrassBlock", "x": 900, "y": 740 },
-    { "type": "GrassBlock", "x": 1100, "y": 740 },
-    { "type": "GrassBlock", "x": 1300, "y": 740 },
-    { "type": "GrassBlock", "x": 1500, "y": 740 },
-    { "type": "GrassBlock", "x": 1600, "y": 740 },
-    { "type": "GrassBlock", "x": 1700, "y": 740 },
-    { "type": "GrassBlock", "x": 1700, "y": 840 },
-    { "type": "GrassBlock", "x": 1800, "y": 840 },
-    { "type": "GrassBlock", "x": 1600, "y": 840 },
-    { "type": "GrassBlock", "x": 1400, "y": 840 },
-    { "type": "GrassBlock", "x": 1300, "y": 840 },
-    { "type": "GrassBlock", "x": 1100, "y": 840 },
-    { "type": "GrassBlock", "x": 1000, "y": 840 },
-    { "type": "GrassBlock", "x": 900, "y": 840 },
-    { "type": "GrassBlock", "x": 1000, "y": 740 },
-    { "type": "GrassBlock", "x": 1200, "y": 840 },
-    { "type": "GrassBlock", "x": 1200, "y": 740 },
-    { "type": "GrassBlock", "x": 1400, "y": 740 },
-    { "type": "GrassBlock", "x": 1500, "y": 840 },
-    { "type": "RampBlockR", "x": 300, "y": 140 },
-    { "type": "RampBlockR", "x": 400, "y": 240 },
-    { "type": "slightRampBlockR", "x": 2100, "y": 240 },
-    { "type": "GrassBlock", "x": 800, "y": 740 },
-    { "type": "GrassBlock", "x": 700, "y": 740 },
-    { "type": "GrassBlock", "x": 700, "y": 840 },
-    { "type": "GrassBlock", "x": 600, "y": 840 },
-    { "type": "RampBlockL", "x": 600, "y": 740 },
-    { "type": "EnemySpawnBlock", "x": 200, "y": 0, "enemyType": "W" },
-    { "type": "EnemySpawnBlock", "x": 600, "y": 200, "enemyType": "R" },
-    { "type": "EnemySpawnBlock", "x": 800, "y": 200, "enemyType": "E" },
-    { "type": "EnemySpawnBlock", "x": 1000, "y": 200, "enemyType": "C" },
-    { "type": "EnemySpawnBlock", "x": 1200, "y": 200, "enemyType": "K" },
-    { "type": "EnemySpawnBlock", "x": 1400, "y": 200, "enemyType": "I" },
-    { "type": "EnemySpawnBlock", "x": 1600, "y": 200, "enemyType": "N" },
-    { "type": "EnemySpawnBlock", "x": 1900, "y": 200, "enemyType": "G" },
-    { "type": "EnemySpawnBlock", "x": 700, "y": 600, "enemyType": "W" },
-    { "type": "EnemySpawnBlock", "x": 1000, "y": 600, "enemyType": "H" },
-    { "type": "EnemySpawnBlock", "x": 1200, "y": 600, "enemyType": "E" },
-    { "type": "EnemySpawnBlock", "x": 1400, "y": 600, "enemyType": "E" },
-    { "type": "EnemySpawnBlock", "x": 1600, "y": 600, "enemyType": "L" },
-    { "type": "GrassBlock", "x": 1800, "y": 740 },
-    { "type": "EnemySpawnBlock", "x": 1900, "y": 600, "enemyType": "S" },
-    { "type": "GrassBlock", "x": 2100, "y": 440 },
-    { "type": "GrassBlock", "x": 2100, "y": 840 },
-    { "type": "GrassBlock", "x": 1900, "y": 840 },
-    { "type": "GrassBlock", "x": 1900, "y": 740 },
-    { "type": "GrassBlock", "x": 2000, "y": 740 },
-    { "type": "GrassBlock", "x": 2000, "y": 840 },
-    { "type": "GrassBlock", "x": 2000, "y": 440 }
-  ],
-  "commands": [],
-  "buildingBlockTypes": [
-    { "name": "Basic Block", "key": "1", "type": "BasicWoodBlock", "limit": 0 }
-  ]
-}
