+++ public/pages/changeLog/changeLog.html
@@ -117,5 +117,5 @@
             </ul>
         </div>
     </div>
-    <script src="js/changeLog.js"></script>
+    <script src="/pages/changeLog/changeLog.js"></script>
 </body>
