+++ public/pages/index/index.html
@@ -39,14 +39,14 @@
     </div>
 </div>
 <div class="action-container">
-    <button class="action-button" id="login">Login</button>
-    <button class="action-button" id="start-game">Start Game</button>
-    <button class="action-button" id="change-log">Change Log <p
+    <a class="action-button" id="login" href ="login">Login</a>
+    <a class="action-button" id="start-game" href="levels">Start Game</a>
+    <a class="action-button" id="change-log" href="changeLog">Change Log <p
             style="font-size: 12px; color: #ff0000; margin: 0; padding: 0;">And vote on new features!</p>
-    </button>
+    </a>
 </div>
 
-<script type="module" src="js/game/index.js"></script>
+<script type="module" src="/pages/index/index.js"></script>
 </body>
 
 </html>
