+++ public/pages/login/login.html
@@ -26,6 +26,6 @@
             </form>
             <div id="message"></div>
         </div>
-        <script type="module" src="login.js"></script>
+        <script type="module" src="/pages/login/login.js"></script>
     </body>
 </html>
