+++ public/pages/login/login.js
@@ -1,4 +1,4 @@
-import LevelHandler from "./js/loaders/levelHandler.js";
+import LevelHandler from "/js/loaders/levelHandler.js";
 // start the level handler (required for syncing data with the server)
 const levelHandler = new LevelHandler();
