+++ public/pages/register/register.js
